@@ -38,7 +38,7 @@ install_requires =
     astropy>=3
     ginga>=2.7
     scipy>=0.18
-python_requires = >=3.5
+python_requires = >=3.7
 
 [options.extras_require]
 test =
@@ -61,7 +61,6 @@ console_scripts =
 source = stginga
 omit =
     stginga/_astropy_init*
-    stginga/conftest*
     stginga/setup_package*
     stginga/*/setup_package*
     stginga/*/*/setup_package*
@@ -70,7 +69,6 @@ omit =
     stginga/*/*/tests/*
     stginga/version*
     */stginga/_astropy_init*
-    */stginga/conftest*
     */stginga/setup_package*
     */stginga/*/setup_package*
     */stginga/*/*/setup_package*
